requirements.txt
openai
pandas
numpy
python-dotenv
jupyter
ipykernel